[Add] separate the parameter check and add unit test
