+++ aceCoder/aceCoder/Add.h
@@ -4,24 +4,11 @@
 
 class Add {
 public:
-	Add(vector<member>& members) : members_(members) {}
+	Add(vector<member>& members);
 	virtual ~Add() = default;
 
-	bool isValid  (const member &newmem) const {
-		for (const auto& mem : members_) {
-			if (mem.employeeNum == newmem.employeeNum)
-				return false;
-		}
-
-		if (newmem.name == "" || newmem.phoneNum == "" || newmem.birthday == 0 )
-			return false;
-		return true;
-	}
-
-	int run(const member& newmem) {
-		members_.push_back(newmem);
-		return 1;
-	}
+	bool isValid(const member& newmem) const;
+	int run(const member& newmem);
 private:
 	vector<member> &members_;
 };
