+++ aceCoder/aceCoder/ParameterChecker.h
@@ -48,87 +48,10 @@ public:
 		}
 		
 		return results;
-	}
-	
-	bool isCMDValid(const string& str) {
-		if (str != "ADD" && str != "DEL" && str != "SCH" && str != "MOD")
-			return false;
-		return true;
-	}
+	}	
 
-	bool isEmpoleeNumValid(const string& str) {
-		if (str.size() > EMPLOYEENUMSIZE)
-			return false;
-		return true;
-	}
-
-	bool isNameValid(const string& str) {
-		if (str.size() > MAXNAMESIZE)
-			return false;
-		return true;
-	}
-
-	bool isCLValid(const string& str) {
-		if (str != "CL1" && str != "CL2" && str != "CL3" && str != "CL4")
-			return false;
-		return true;
-	}
-
-	bool isPhoneNumValid(const string& str) {
-		if (str.size() > PHONNUBERSIZE || str.substr(0, 3) != "010"|| str.substr(3, 1) != "-" || str.substr(3, 1) != "-" || str.substr(8, 1) != "-")
-			return false;
-			
-		return true;
-	}
-
-	bool isBirthdayValid(const string& str) {
-		if (str.size() > BIRTHDAYSIZE)
-			return false;
-		return true;
-	}
-
-	bool isCELTIValid(const string& str) {
-		if (str != "ADV" && str != "PRO" && str != "EXPERT")
-			return false;
-		return true;
-	}
-
-	bool isValid(const vector<string>& values)  {
-		if (values.size() != PARAINDEX::PARAINDEXEND)
-			return false;
-
-		for (int i = 0; i < values.size(); i++) {
-			if (i == PARAINDEX::IDXCMD) {
-				if (isCMDValid(values[i]) == false)
-					return false;					
-			}
-			else if (i == PARAINDEX::IDXEMPLOYEENUM) {
-				if (isEmpoleeNumValid(values[i]) == false)
-					return false;
-			}
-			else if (i == PARAINDEX::IDXNAME) {
-				if (isNameValid(values[i]) == false)
-					return false;
-			}
-			else if (i == PARAINDEX::IDXCL) {
-				if (isCLValid(values[i]) == false)
-					return false;
-			}
-			else if (i == PARAINDEX::IDXPHONENUM) {
-				if (isPhoneNumValid(values[i]) == false)
-					return false;
-			}
-			else if (i == PARAINDEX::IDXBIRTHDAY) {
-				if (isBirthdayValid(values[i]) == false)
-					return false;
-			}
-			else if (i == PARAINDEX::IDXCERTI) {
-				if (isCELTIValid(values[i]) == false)
-					return false;
-			}
-		}
-
-		return true;
+	virtual bool isValid(const vector<string>& values)  {		
+		return false;
 	}
 
 	CL getCL(string& str) const {
@@ -150,6 +73,4 @@ public:
 
 		return CERTI::EX;
 	}
-private:
-
 };
