+++ aceCoder/aceCoder/test_ADD.cpp
@@ -1,11 +1,11 @@
 #pragma once
 #include "pch.h"
 #include"Add.h"
-#include "ParameterChecker.h"
+#include "ADDParmeterChecker.h"
 
 TEST(ADD_TESTSUITE, TEST_PARAMETER) {
 	string str = "ADD, , , ,18050301,KYUMOK KIM,CL2,010-9777-6055,19980906,PRO";
-	ParameterChecker* para = new ParameterChecker();		
+	ParameterChecker* para = new AddParmeterChecker();
 	EXPECT_TRUE(para->isValid(para->parse(str)));	
 	str = "XXX, , , ,18050301,KYUMOK KIM,CL2,010-9777-6055,19980906";
 	EXPECT_FALSE(para->isValid(para->parse(str)));
@@ -35,10 +35,9 @@ TEST(ADD_TESTSUITE, TEST_VALID) {
 	delete add;
 }
 
-
 TEST(ADD_TESTSUITE, TEST_INPUTFILE) {
 	vector<member> members;
-	ParameterChecker* para = new ParameterChecker();
+	ParameterChecker* para = new AddParmeterChecker();
 	vector<string>lines = para->loadTxt(".\\input\\inputAdd.txt");
 	ASSERT_NE(0, lines.size());
 	Add* add = new Add(members);
@@ -49,6 +48,33 @@ TEST(ADD_TESTSUITE, TEST_INPUTFILE) {
 		EXPECT_TRUE(add->isValid(mem));
 		EXPECT_EQ(1, add->run(mem));
 	}
+
+	delete add;
+	delete para;
+}
+
+TEST(ADD_TESTSUITE, TEST_INPUTFILE2) {
+	vector<member> members;
+	ParameterChecker* para = new AddParmeterChecker();
+	vector<string>lines = para->loadTxt(".\\input\\input_20_20.txt");
+	ASSERT_NE(0, lines.size());
+	Add* add = new Add(members);
+	for (const auto& line : lines) {
+		vector<string> values = para->parse(line);
+		if (values[IDXCMD] == "ADD")
+		{
+			if (!para->isValid(values))
+				continue;
+			const member mem = para->convert(values);
+			if (!add->isValid(mem))
+				continue;
+			add->run(mem);
+		}
+		else
+			continue;
+	}
+
+	EXPECT_EQ(20, members.size());
 
 	delete add;
 	delete para;
